requirements.txt
streamlit
snowflake-connector-python
snowflake-connector-python[pandas]